warehouse_management/requirements.txt
astroid==1.5.3
autopep8==1.3.5
coverage==4.4.2
Django==2.0.1
django-bootstrap4==0.0.4
djangorestframework==3.8.2
Faker==0.8.8
flake8==3.5.0
greenlet==0.4.12
isort==4.2.15
jedi==0.11.1
json-rpc==1.10.8
lazy-object-proxy==1.3.1
mccabe==0.6.1
mock==2.0.0
msgpack-python==0.4.8
neovim==0.2.0
parso==0.1.1
pbr==4.0.2
pep8==1.7.1
pycodestyle==2.3.1
pyflakes==1.6.0
Pygments==2.2.0
pylint==1.7.4
python-dateutil==2.6.1
pytz==2017.3
redgreenunittest==0.1.1
rope==0.10.7
service-factory==0.1.5
six==1.11.0
text-unidecode==1.1
wrapt==1.10.11
yapf==0.21.0
